+++ src/main/java/com/imsweb/algorithms/survival/SurvivalTimeInputRecordDto.java
@@ -3,7 +3,9 @@
  */
 package com.imsweb.algorithms.survival;
 
-public class SurvivalTimeInputRecordDto {
+import org.apache.commons.lang3.math.NumberUtils;
+
+public class SurvivalTimeInputRecordDto implements Comparable<SurvivalTimeInputRecordDto> {
 
     private String _patientIdNumber;
 
@@ -30,7 +32,9 @@ public class SurvivalTimeInputRecordDto {
     private String _sequenceNumberCentral;
 
     private String _typeOfReportingSource;
-   
+
+    private int _sortedIndex;
+
     public String getPatientIdNumber() {
         return _patientIdNumber;
     }
@@ -135,4 +139,49 @@ public class SurvivalTimeInputRecordDto {
         _typeOfReportingSource = typeOfReportingSource;
     }
 
+    public int getSortedIndex() {
+        return _sortedIndex;
+    }
+
+    public void setSortedIndex(int sortedIndex) {
+        _sortedIndex = sortedIndex;
+    }
+
+    @Override
+    public int compareTo(SurvivalTimeInputRecordDto other) {
+        int year = NumberUtils.isDigits(_dateOfDiagnosisYear) ? Integer.parseInt(_dateOfDiagnosisYear) : 9999;
+        int month = NumberUtils.isDigits(_dateOfDiagnosisMonth) ? Integer.parseInt(_dateOfDiagnosisMonth) : 99;
+        int day = NumberUtils.isDigits(_dateOfDiagnosisDay) ? Integer.parseInt(_dateOfDiagnosisDay) : 99;
+        if (month == 99)
+            day = 99;
+        int seqNum = NumberUtils.isDigits(_sequenceNumberCentral) ? Integer.parseInt(_sequenceNumberCentral) : -1;
+        if (seqNum >= 60 && seqNum <= 97)
+            seqNum = seqNum + 100;
+
+        int otherYear = NumberUtils.isDigits(other._dateOfDiagnosisYear) ? Integer.parseInt(other._dateOfDiagnosisYear) : 9999;
+        int otherMonth = NumberUtils.isDigits(other._dateOfDiagnosisMonth) ? Integer.parseInt(other._dateOfDiagnosisMonth) : 99;
+        int otherDay = NumberUtils.isDigits(other._dateOfDiagnosisDay) ? Integer.parseInt(other._dateOfDiagnosisDay) : 99;
+        if (otherMonth == 99)
+            otherDay = 99;
+        int otherSeqNum = NumberUtils.isDigits(other._sequenceNumberCentral) ? Integer.parseInt(other._sequenceNumberCentral) : -1;
+        if (otherSeqNum >= 60 && otherSeqNum <= 97)
+            otherSeqNum = otherSeqNum + 100;
+
+        if (year == 9999 || otherYear == 9999)
+            return seqNum - otherSeqNum;
+        else if (year != otherYear)
+            return year - otherYear;
+        else {
+            if (month == 99 || otherMonth == 99)
+                return seqNum - otherSeqNum;
+            else if (month != otherMonth)
+                return month - otherMonth;
+            else {
+                if (day == 99 || otherDay == 99 || day == otherDay)
+                    return seqNum - otherSeqNum;
+                else
+                    return day - otherDay;
+            }
+        }
+    }
 }